# api/docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.CreateIncomeProduct:
    properties:
      barcode:
        type: string
      category_id:
        type: string
      income_id:
        type: string
      income_price:
        type: number
      product_name:
        type: string
      quantity:
        type: integer
    type: object
...
  models.IncomeProduct:
    properties:
      barcode:
        type: string
      category_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      income_id:
        type: string
      income_price:
        type: number
      product_name:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.UpdateIncomeProduct:
    properties:
      barcode:
        type: string
      category_id:
        type: string
      id:
        type: string
      income_id:
        type: string
      income_price:
        type: number
      product_name:
        type: string
      quantity:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /v1/income_product:
    get:
      consumes:
      - application/json
      description: Get a list of income products with optional filtering.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Number of items to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of items to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of income products
          schema:
            items:
              $ref: '#/definitions/models.IncomeProduct'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a list of income products
      tags:
      - income_product
    post:
      consumes:
      - application/json
      description: Create a new income product in the market system.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Income product information
        in: body
        name: income_product
        required: true
        schema:
          $ref: '#/definitions/models.CreateIncomeProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created income product
          schema:
            $ref: '#/definitions/models.IncomeProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new income product
      tags:
      - income_product
  /v1/income_product/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing income product.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Income Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete an income product
      tags:
      - income_product
    get:
      consumes:
      - application/json
      description: Get income product details by its ID.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Income Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Income product details
          schema:
            $ref: '#/definitions/models.IncomeProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Income product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get an income product by ID
      tags:
      - income_product
    put:
      consumes:
      - application/json
      description: Update an existing income product.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Income Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated income product information
        in: body
        name: income_product
        required: true
        schema:
          $ref: '#/definitions/models.UpdateIncomeProduct'
      produces:
      - application/json
      responses:
        "202":
          description: Updated income product
          schema:
            $ref: '#/definitions/models.IncomeProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Income product not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update an income product
      tags:
      - income_product
...
swagger: "2.0"